pyproject.toml
[tool.poetry]
name = "pype"
version = "0.1.0"
description = ""
authors = ["German Arutyunov <germanarutyunov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
hydra-slayer = "^0.5.0"
pyyaml = "^6.0.2"
python-gitlab = "^5.0.0"
pygithub = "^2.5.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.3"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
